+++ CHANGELOG.md
@@ -1,10 +1,16 @@
+# version 0.0.2
+
+- Version number,
+- Implement basic configuration (folder where data is to be stored)
+
 # version 0.0.1
 
-Created repo.
+Created Github repo.
 
 # Unreleased
 
-[ ] Decide about cli interface (dialoguer?)\
+[ ] make .edit_config suggest `$suv_folder/suv_archive` as default for an archive
 [ ] Implement `add_thesis`
+[ ] Decide about cli interface (dialoguer?)\
 
 # History